FIX/Reacomodar direcciones de variables dsp de modularizar
