+++ client/src/components/CardContainer/CardContainer.jsx
@@ -3,7 +3,7 @@ import { useDispatch, useSelector } from "react-redux";
 import {
   getAllAnimals,
   setUserAnimalsToLoading,
-} from "../../redux/actions/actions";
+} from "../../redux/actions/animal-actions/animal-actions";
 import loading from "../../assets/loading.gif";
 import { Pagination } from "../Pagination/Pagination";
 import { PropsColumns } from "../PropsColumns/PropsColumns";
